Test memory mode in `io.fs` (#8468)

+++ test/io/test_fs.py
@@ -2,8 +2,19 @@ import zipfile
 from os import path as osp
 
 import fsspec
+import pytest
 
+from torch_geometric.data import extract_zip
 from torch_geometric.io import fs
+
+
+@pytest.fixture(params=['file', 'memory'])
+def tmp_fs_path(request, tmp_path) -> str:
+    if request.param == 'file':
+        return tmp_path.resolve().as_posix()
+    elif request.param == 'memory':
+        return f'memory://{tmp_path}'
+    raise NotImplementedError
 
 
 def test_get_fs():
@@ -16,41 +27,39 @@ def test_normpath():
     assert fs.normpath('memory:////home') == 'memory:////home'
 
 
-def test_exists(tmp_path):
-    path = osp.join(tmp_path, 'file.txt')
+def test_exists(tmp_fs_path):
+    path = osp.join(tmp_fs_path, 'file.txt')
     assert not fs.exists(path)
     with fsspec.open(path, 'w') as f:
         f.write('here')
     assert fs.exists(path)
 
 
-def test_makedirs(tmp_path):
-    path = osp.join(tmp_path, '1', '2')
+def test_makedirs(tmp_fs_path):
+    path = osp.join(tmp_fs_path, '1', '2')
     assert not fs.isdir(path)
     fs.makedirs(path)
     assert fs.isdir(path)
 
 
-def test_ls(tmp_path):
-    tmp_path = tmp_path.resolve().as_posix()
-
+def test_ls(tmp_fs_path):
     for i in range(2):
-        with fsspec.open(osp.join(tmp_path, str(i)), 'w') as f:
+        with fsspec.open(osp.join(tmp_fs_path, str(i)), 'w') as f:
             f.write('here')
-    res = fs.ls(tmp_path)
+    res = fs.ls(tmp_fs_path)
     assert len(res) == 2
-    expected_protocol = fs.get_fs(tmp_path).protocol
+    expected_protocol = fs.get_fs(tmp_fs_path).protocol
     assert all(fs.get_fs(path).protocol == expected_protocol for path in res)
 
 
-def test_cp(tmp_path):
-    src = osp.join(tmp_path, 'src')
+def test_cp(tmp_fs_path):
+    src = osp.join(tmp_fs_path, 'src')
     for i in range(2):
         with fsspec.open(osp.join(src, str(i)), 'w') as f:
             f.write('here')
     assert fs.exists(src)
 
-    dst = osp.join(tmp_path, 'dst')
+    dst = osp.join(tmp_fs_path, 'dst')
     assert not fs.exists(dst)
 
     # Can copy a file to new name:
@@ -70,24 +79,29 @@ def test_cp(tmp_path):
         fs.exists(osp.join(dst, str(i)))
 
 
-def test_cp_extract(tmp_path):
+def test_extract(tmp_fs_path):
     def make_zip(path: str):
         with fsspec.open(path, mode='wb') as f:
             with zipfile.ZipFile(f, mode='w') as z:
                 z.writestr('1', b'data')
                 z.writestr('2', b'data')
 
-    src = osp.join(tmp_path, 'src', 'test.zip')
+    src = osp.join(tmp_fs_path, 'src', 'test.zip')
     make_zip(src)
     assert len(fsspec.open_files(f'zip://*::{src}')) == 2
 
-    dst = osp.join(tmp_path, 'dst')
+    dst = osp.join(tmp_fs_path, 'dst')
     assert not fs.exists(dst)
 
-    # Can copy:
-    fs.cp(src, dst)
-    assert fs.exists(dst)
-    fs.rm(dst)
+    # Can copy and extract afterwards:
+    if fs.isdisk(tmp_fs_path):
+        fs.cp(src, osp.join(dst, 'test.zip'))
+        assert fs.exists(osp.join(dst, 'test.zip'))
+        extract_zip(osp.join(dst, 'test.zip'), dst)
+        assert len(fs.ls(dst)) == 3
+        for i in range(2):
+            fs.exists(osp.join(dst, str(i)))
+        fs.rm(dst)
 
     # Can copy and extract:
     fs.cp(src, dst, extract=True)
